Create N_ary_tree.py
First, we have to convert the list of edges into a proper dictionary which will indicate the tree structure of the given edges using the function "createDict()". Then we will make it into a tree structure using the function   "createNode()". Finally, we print the level order traversal of the tree
1   
2  3   
4  5  8  7   
6  9